+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added API endpoint for removing identity proofs.
+
 ### Changed
 
 - Copied custom fields from Instance API v1 response to v2 response.